Refactoring & Cleanup
