+++ src/encoded/static/components/item-pages/CaseView/VariantSampleListController.js
@@ -142,6 +142,7 @@ export class VariantSampleListController extends React.PureComponent {
                         "variant_samples.variant_sample_item.uuid",
                         "variant_samples.variant_sample_item.display_title",
                         "variant_samples.variant_sample_item.finding_table_tag",
+                        "variant_samples.variant_sample_item.actions",
                         "variant_samples.variant_sample_item.associated_genotype_labels.proband_genotype_label",
                         "variant_samples.variant_sample_item.associated_genotype_labels.mother_genotype_label",
                         "variant_samples.variant_sample_item.associated_genotype_labels.father_genotype_label",
