+++ src/encoded/static/components/navigation/components/index.js
@@ -1,6 +1,5 @@
 'use strict';
 
-export { SearchBar } from './SearchBar';
 export { TestWarning } from './TestWarning';
 export { LoginNavItem, NotLoggedInAlert } from './LoginNavItem';
 export { CollapsedNav } from './CollapsedNav';
