+++ src/encoded/types/case.py
@@ -280,9 +280,10 @@ def _build_case_embedded_list():
         "report.last_modified.*",
         "report.status",
         "report.accession",
-        "report.case.accession",
+        "report.case.accession", # This is same as this Item's accession, no?
+        "report.variant_samples.uuid",
 
-        # FilterSet LinkTo
+        # FilterSet linkTo
         "active_filterset.@id",
         "active_filterset.title",
